requirements.txt
antlr4-python3-runtime>=4.9.3
jinja2

# test requirements
pre-commit
black
mypy
flake8
autopep8
pylint
pytest
pytest-cov
pytest-mock
pytest-spark
pytest-rerunfailures
sphinx>=2.4.0
sphinx-rtd-theme
sphinx-autodoc-typehints
psutil

# publish to pypi
wheel
twine

cibuildwheel==2.1.3
speedy-antlr-tool
